Rename common/system to common/platform
