+++ CMakeLists.txt
@@ -219,19 +219,19 @@ set(LIGHTSTEP_SRCS src/common/utility.cpp
                    )
 
 if (WIN32)
-  list(APPEND LIGHTSTEP_SRCS src/common/system/error_windows.cpp
-                             src/common/system/network_windows.cpp
-                             src/common/system/network_environment_windows.cpp
-                             src/common/system/utility_windows.cpp
-                             src/common/system/string_windows.cpp
-                             src/common/system/fork_windows.cpp)
+  list(APPEND LIGHTSTEP_SRCS src/common/platform/error_windows.cpp
+                             src/common/platform/network_windows.cpp
+                             src/common/platform/network_environment_windows.cpp
+                             src/common/platform/utility_windows.cpp
+                             src/common/platform/string_windows.cpp
+                             src/common/platform/fork_windows.cpp)
 else()
-  list(APPEND LIGHTSTEP_SRCS src/common/system/error_unix.cpp
-                             src/common/system/network_unix.cpp
-                             src/common/system/network_environment_unix.cpp
-                             src/common/system/utility_unix.cpp
-                             src/common/system/string_unix.cpp
-                             src/common/system/fork_unix.cpp)
+  list(APPEND LIGHTSTEP_SRCS src/common/platform/error_unix.cpp
+                             src/common/platform/network_unix.cpp
+                             src/common/platform/network_environment_unix.cpp
+                             src/common/platform/utility_unix.cpp
+                             src/common/platform/string_unix.cpp
+                             src/common/platform/fork_unix.cpp)
 endif()
 
 if (WITH_GRPC)
