+++ src/common/platform/BUILD
