+++ src/common/platform/error.h
