+++ src/common/platform/memory.h
