+++ src/common/platform/network.h
