+++ src/common/platform/network_environment.h
