+++ src/common/platform/network_windows.cpp
@@ -1,61 +1,59 @@
-#include "common/system/network.h"
+#include "common/platform/network.h"
 
-#include <climits>
+#include <sstream>
+#include <stdexcept>
 
-#include <unistd.h>
-#include <fcntl.h>
-#include <sys/socket.h>
-#include <sys/types.h>
-#include <sys/uio.h>
+#include <winsock2.h>
 
 namespace lightstep {
-//--------------------------------------------------------------------------------------------------
-// IoVecMax
-//--------------------------------------------------------------------------------------------------
-const int IoVecMax = IOV_MAX;
-
 //--------------------------------------------------------------------------------------------------
 // MakeIoVec
 //--------------------------------------------------------------------------------------------------
 IoVec MakeIoVec(void* data, size_t length) noexcept {
-  return IoVec{data, length};
+  return IoVec{static_cast<ULONG>(length), static_cast<char*>(data)};
 }
 
 //--------------------------------------------------------------------------------------------------
 // Read
 //--------------------------------------------------------------------------------------------------
 int Read(FileDescriptor socket, void* data, size_t size) noexcept {
-  return static_cast<int>(::read(socket, data, size));
+  return ::recv(socket, static_cast<char*>(data), static_cast<int>(size), 0);
 }
 
 //--------------------------------------------------------------------------------------------------
 // WriteV
 //--------------------------------------------------------------------------------------------------
 int WriteV(FileDescriptor socket, const IoVec* iov, int iovcnt) noexcept {
-  return ::writev(socket, iov, iovcnt);
+  DWORD num_bytes_sent;
+  auto rcode = ::WSASend(socket, const_cast<IoVec*>(iov), iovcnt, &num_bytes_sent, 0,
+                         nullptr, nullptr);
+  if (rcode != 0) {
+    return rcode;
+  }
+  return static_cast<int>(num_bytes_sent);
 }
 
 //--------------------------------------------------------------------------------------------------
 // SetSocketNonblocking
 //--------------------------------------------------------------------------------------------------
 int SetSocketNonblocking(FileDescriptor socket) noexcept {
-  return ::fcntl(socket, F_SETFL,
-                      ::fcntl(socket, F_GETFL, 0) | O_NONBLOCK);
+  u_long nonblocking_mode = 1;
+  return ::ioctlsocket(socket, FIONBIO, &nonblocking_mode);
 }
 
 //--------------------------------------------------------------------------------------------------
 // SetSocketReuseAddress
 //--------------------------------------------------------------------------------------------------
 int SetSocketReuseAddress(FileDescriptor socket) noexcept {
-  int optvalue = 1;
+  BOOL optvalue = TRUE;
   return ::setsockopt(socket, SOL_SOCKET, SO_REUSEADDR,
-                      static_cast<void*>(&optvalue), sizeof(int));
+                      reinterpret_cast<char*>(&optvalue), sizeof(optvalue));
 }
 
 //--------------------------------------------------------------------------------------------------
 // CloseSocket
 //--------------------------------------------------------------------------------------------------
 int CloseSocket(FileDescriptor socket) noexcept {
-  return ::close(socket);
+  return ::closesocket(socket);
 }
 } // namespace lightstep
