+++ src/common/platform/string.h
