+++ src/common/platform/string_unix.cpp
@@ -1,9 +1,9 @@
-#include "common/system/string.h"
+#include "common/platform/string.h"
 
-#include <string.h>
+#include <strings.h>
 
 namespace lightstep {
 int StrCaseCmp(const char* s1, const char* s2) noexcept {
-  return ::_stricmp(s1, s2);
+  return ::strcasecmp(s1, s2);
 }
 } // namespace lightstep
