+++ src/common/platform/utility.h
