+++ src/common/platform/utility_windows.cpp
@@ -1,4 +1,4 @@
-#include "common/system/utility.h"
+#include "common/platform/utility.h"
 
 #include <memory>
 #include <string>
