+++ src/network/socket.cpp
@@ -1,6 +1,6 @@
 #include "network/socket.h"
 
-#include "common/system/error.h"
+#include "common/platform/error.h"
 
 #include <cassert>
 #include <cstring>
