+++ src/recorder/BUILD
@@ -49,7 +49,7 @@ lightstep_cc_library(
         "fork_aware_recorder.cpp",
     ],
     deps = [
-        "//src/common/system:fork_lib",
+        "//src/common/platform:fork_lib",
         ":recorder_interface",
     ],
 )
