+++ src/recorder/stream_recorder/BUILD
@@ -96,7 +96,7 @@ lightstep_cc_library(
         "//src/common:noncopyable_lib",
         "//src/common:protobuf_lib",
         "//src/common:chunk_circular_buffer_lib",
-        "//src/common/system:network_environment_lib",
+        "//src/common/platform:network_environment_lib",
         "//src/network:event_lib",
         "//src/network:timer_event_lib",
         "//src/recorder:fork_aware_recorder_lib",
@@ -154,8 +154,8 @@ lightstep_cc_library(
     ],
     deps = [
         ":satellite_endpoint_manager_lib",
-        "//src/common/system:error_lib",
-        "//src/common/system:network_lib",
+        "//src/common/platform:error_lib",
+        "//src/common/platform:network_lib",
         "//src/common:chunk_circular_buffer_lib",
         "//src/common:noncopyable_lib",
         "//src/common:random_lib",
@@ -186,7 +186,7 @@ lightstep_cc_library(
         "//include/lightstep:tracer_interface",
         "//lightstep-tracer-common:collector_proto_cc",
         "//src/common:protobuf_lib",
-        "//src/common/system:string_lib",
+        "//src/common/platform:string_lib",
         "//src/common:utility_lib",
         "//src/common:circular_buffer_lib",
         "//src/network:fragment_array_input_stream_lib",
