+++ src/recorder/stream_recorder/satellite_connection.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common/system/network.h"
+#include "common/platform/network.h"
 #include "common/noncopyable.h"
 #include "network/event.h"
 #include "network/socket.h"
