+++ src/recorder/stream_recorder/stream_recorder.h
@@ -11,7 +11,7 @@
 #include "common/chunk_circular_buffer.h"
 #include "common/logger.h"
 #include "common/noncopyable.h"
-#include "common/system/network_environment.h"
+#include "common/platform/network_environment.h"
 #include "lightstep/tracer.h"
 #include "network/event_base.h"
 #include "network/timer_event.h"
